requirements.txt
streamlit
streamlit_chat
llama_index
langchain_community
opencv-python
pillow
ultralytics